+++ Infrastructure/Repository/Interfaces/IThresholdRepository.cs
@@ -5,10 +5,10 @@ namespace Infrastructure.Repository.Interfaces
 {
     public interface IThresholdRepository
     {
-        Task<Threshold> GetThresholdByIdAsync(int thresholdId);
+        Task<Threshold> GetThresholdByIdAsync(long thresholdId);
         Task<List<Threshold>> GetAllThresholdsAsync();
         Task AddThresholdAsync(Threshold threshold);
         Task UpdateThresholdAsync(Threshold threshold);
-        Task DeleteThresholdAsync(int thresholdId);
+        Task DeleteThresholdAsync(long thresholdId);
     }
 }
